LayoutTests/media/video-src-change.html

<video src=bogus controls></video>
<div>
Test that changing src attribute triggers load when network state is empty.<br>
Test that it does not trigger load when network state is non-empty.
</div>
<script src=video-test.js></script>
<script>
    waitForEvent('emptied', function () {
        testExpected("video.networkState", HTMLMediaElement.EMPTY);
        video.setAttribute('src', 'content/test.mp4')
        waitForEvent("loadstart", function () { testExpected("relativeURL(video.currentSrc)", "content/test.mp4"); });

        waitForEvent('load', function () {
            testExpected("video.networkState", HTMLMediaElement.EMPTY, "!=");
            video.setAttribute('src', 'content/test.wav');

            waitForEvent('loadstart', function () { 
                logResult("LOADSTART fired");
                endTest();
            } );

            setTimeout(endTest, 500);
        });
    });
</script>
